Image Converter/requirements.txt
Pillow
customtkinter